Other: Extract `enableModelIfOneIsEnabled()` to own file.
