+++ tests/dropdown/utils.js
@@ -18,8 +18,7 @@ import createDropdownView from '../../src/dropdown/helpers/createdropdownview';
 
 import {
 	addListViewToDropdown,
-	addToolbarToDropdown,
-	enableModelIfOneIsEnabled
+	addToolbarToDropdown
 } from '../../src/dropdown/utils';
 
 describe( 'utils', () => {
@@ -36,31 +35,6 @@ describe( 'utils', () => {
 		if ( dropdownView.element ) {
 			dropdownView.element.remove();
 		}
-	} );
-
-	describe( 'enableModelIfOneIsEnabled()', () => {
-		it( 'Bind to #isEnabled of each observable  and set it true if any observable #isEnabled is true', () => {
-			const observables = [
-				new Model( { isEnabled: false } ),
-				new Model( { isEnabled: false } ),
-				new Model( { isEnabled: false } )
-			];
-			enableModelIfOneIsEnabled( model, observables );
-
-			expect( model.isEnabled ).to.be.false;
-
-			observables[ 0 ].isEnabled = true;
-
-			expect( model.isEnabled ).to.be.true;
-
-			observables[ 0 ].isEnabled = false;
-
-			expect( model.isEnabled ).to.be.false;
-
-			observables[ 1 ].isEnabled = true;
-
-			expect( model.isEnabled ).to.be.true;
-		} );
 	} );
 
 	describe( 'addListViewToDropdown()', () => {
